sophia: add range queue data structure
This very simple data structure will be used
instead of red-black tree for fast
planner updates (delete + insert).

It maps requested value into a
sequential list of predefined ranges
(with specified step) and associated queue list.
Key precision can vary +-step.

Data structure has O(1) for add/delete and
iteration.

+++ rt/sr_cmp.c
@@ -44,14 +44,14 @@ sr_cmppointer_of(char *name)
 	name += 8;
 	errno = 0;
 	char *end;
-	uintptr_t pointer = strtoull(name, &end, 16);
+	unsigned long long int pointer = strtoull(name, &end, 16);
 	if (pointer == 0 && end == name) {
 		return NULL;
 	} else
 	if (pointer == ULLONG_MAX && errno) {
 		return NULL;
 	}
-	return (void*)pointer;
+	return (void*)(uintptr_t)pointer;
 }
 
 int sr_cmpset(srcomparator *c, char *name)
